requirements.txt
numpy==2.2.6
Pillow==12.0.0
Requests==2.32.5
selenium==4.39.0
webdriver_manager==4.0.2